refactor: 패키지 구조 client 별로 구분
