+++ src/main/kotlin/dnd11th/blooming/api/service/user/oauth/IdTokenProcessor.kt
@@ -2,7 +2,7 @@ package dnd11th.blooming.api.service.user.oauth
 
 import com.fasterxml.jackson.databind.ObjectMapper
 import com.fasterxml.jackson.module.kotlin.readValue
-import dnd11th.blooming.client.dto.OidcPublicKeys
+import dnd11th.blooming.client.oauth.OidcPublicKeys
 import dnd11th.blooming.domain.entity.user.OidcUser
 import io.jsonwebtoken.Claims
 import org.springframework.stereotype.Component
