+++ src/main/kotlin/dnd11th/blooming/api/service/user/oauth/IdTokenValidator.kt
@@ -1,6 +1,6 @@
 package dnd11th.blooming.api.service.user.oauth
 
-import dnd11th.blooming.client.dto.OidcPublicKeys
+import dnd11th.blooming.client.oauth.OidcPublicKeys
 import dnd11th.blooming.common.exception.ErrorType
 import dnd11th.blooming.common.exception.UnAuthorizedException
 import io.jsonwebtoken.Claims
