+++ src/main/kotlin/dnd11th/blooming/api/service/user/oauth/KakaoIdTokenResolver.kt
@@ -1,8 +1,8 @@
 package dnd11th.blooming.api.service.user.oauth
 
 import dnd11th.blooming.api.service.user.oauth.property.KakaoOauthProperties
-import dnd11th.blooming.client.dto.OidcPublicKeys
 import dnd11th.blooming.client.oauth.KakaoOauthClient
+import dnd11th.blooming.client.oauth.OidcPublicKeys
 import dnd11th.blooming.common.exception.ErrorType
 import dnd11th.blooming.common.exception.UnAuthorizedException
 import dnd11th.blooming.domain.entity.user.OidcUser
