+++ src/main/kotlin/dnd11th/blooming/client/oauth/OidcPublicKey.kt
@@ -1,4 +1,4 @@
-package dnd11th.blooming.client.dto
+package dnd11th.blooming.client.oauth
 
 data class OidcPublicKey(
     val kty: String,
