+++ src/main/kotlin/dnd11th/blooming/client/oauth/OidcPublicKeys.kt
@@ -1,4 +1,4 @@
-package dnd11th.blooming.client.dto
+package dnd11th.blooming.client.oauth
 
 data class OidcPublicKeys(
     val keys: List<OidcPublicKey>,
